Struct dungeon_generator::BSPGenerator

pub struct BSPGenerator {
    // some fields omitted
}

Simple Binary Space Partioning dungeon generation algorithm:

Workspace splitting to random rectangles with the help of binary tree

In every rectangle generating a random room

Coridors generating with the help of the binary tree, generated at the first step

Methods

impl BSPGenerator

fn bs_max_deep(self, value: u32) -> Self

levels of binary tree

fn deny_split_deep(self, value: u32) -> Self

level of tree from which node can deny split

fn split_chance_dec(self, value: u32) -> Self

split chance decreasment per tree level starting "deny_split_deep"

fn split_max_coeff(self, value: u32) -> Self

max splitting coefficient

fn split_default_chance(self, value: u32) -> Self

default splitting chance

fn min_room_splittable_size(self, value: u32) -> Self

minimum splittable size of subdungeon

fn coridor_width(self, value: u32) -> Self

constant width of corridors